requirements.txt
streamlit
flask
requests
pymilvus
openai
python-dotenv
beautifulsoup4